feat: split advanced tab into advanced and privacy & security tabs
