+++ apps/web/public/static/locales/en/common.json
@@ -1834,6 +1834,8 @@
   "event_limit_tab_description": "How often you can be booked",
   "event_advanced_tab_description": "Calendar settings & more...",
   "event_advanced_tab_title": "Advanced",
+  "event_privacy_tab_description": "Confirmation, cancellation & privacy",
+  "event_privacy_tab_title": "Privacy & Security",
   "event_payments_tab_description": "Set up payments for event",
   "event_setup_multiple_duration_error": "Event Setup: Multiple durations requires at least 1 option.",
   "event_setup_multiple_duration_default_error": "Event Setup: Please select a valid default duration.",
