+++ packages/features/eventtypes/lib/types.ts
@@ -451,6 +451,7 @@ export type TabMap = {
   availability: React.ReactNode;
   instant?: React.ReactNode;
   limits: React.ReactNode;
+  privacy: React.ReactNode;
   recurring: React.ReactNode;
   setup: React.ReactNode;
   team?: React.ReactNode;
